refactor: restructure HTTP server into a subsystem and extract job domain
The HTTP API lived in a single 2026-line `cmd/server.rs` that mixed request/response types, API-key auth, CORS, JSONP, background job orchestration, and every axum handler. This reshapes it in three steps, all behavior-preserving:

1. Split into cohesive submodules (types/auth/cors/response/util/job_runner/handlers) behind a `pub(crate)` re-export prelude in `mod.rs`, so sibling modules and the existing test module keep resolving names through `use super::*` with zero test edits.

2. Promote the server from `cmd/server/` to top-level `src/server/`. Under `cmd/` each entry is one CLI subcommand; a persistent server is a subsystem, and the MCP server already lived top-level (`main.rs` dispatches `Commands::Mcp => mcp::run_mcp_server()`). `cmd/` is now all subcommands; `main.rs` dispatches `Commands::Server => server::run_server`.

3. Extract the shared job domain — Job, JobProgress, JobStatus, retention/limit constants, and the reachability probe — out of `cmd/job.rs` + `cmd/mod.rs` into top-level `src/job/`. Both `server` and `mcp` now depend on `crate::job` instead of a subsystem reaching into `crate::cmd`. `error_codes` stays in `cmd`.

No logic changes. cargo build + clippy --all-targets clean; full lib suite 1275 passing.
