+++ src/cmd/mod.rs
@@ -1,13 +1,8 @@
 pub mod file;
-pub mod job;
 pub mod payload;
 pub mod pipe;
 pub mod scan;
-pub mod server;
 pub mod url;
-
-use serde::{Deserialize, Serialize};
-use std::fmt;
 
 /// Application-level error codes shared across CLI, REST API, and MCP interfaces.
 /// These codes appear in JSON output (`error.code`, `target_summary.error_code`,
@@ -43,76 +38,9 @@ pub mod error_codes {
     pub const TRUNCATED_PER_HOST_CAP: &str = "TRUNCATED_PER_HOST_CAP";
 }
 
-/// Status of an asynchronous scan job (used by both REST server and MCP).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum JobStatus {
-    Queued,
-    Running,
-    Done,
-    Error,
-    Cancelled,
-}
-
-impl fmt::Display for JobStatus {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Queued => write!(f, "queued"),
-            Self::Running => write!(f, "running"),
-            Self::Done => write!(f, "done"),
-            Self::Error => write!(f, "error"),
-            Self::Cancelled => write!(f, "cancelled"),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn job_status_display_matches_lowercase_variant_name() {
-        assert_eq!(JobStatus::Queued.to_string(), "queued");
-        assert_eq!(JobStatus::Running.to_string(), "running");
-        assert_eq!(JobStatus::Done.to_string(), "done");
-        assert_eq!(JobStatus::Error.to_string(), "error");
-        assert_eq!(JobStatus::Cancelled.to_string(), "cancelled");
-    }
-
-    /// The Display impl and the `#[serde(rename_all = "lowercase")]`
-    /// representation must agree — REST and MCP clients parse the JSON
-    /// form, and CLI logs print the Display form. Drift between them
-    /// would silently break consumers that compare the two strings.
-    #[test]
-    fn job_status_serde_matches_display() {
-        let variants = [
-            JobStatus::Queued,
-            JobStatus::Running,
-            JobStatus::Done,
-            JobStatus::Error,
-            JobStatus::Cancelled,
-        ];
-        for v in variants {
-            let json = serde_json::to_string(&v).unwrap();
-            // serde_json wraps the variant name in quotes.
-            assert_eq!(json, format!("\"{}\"", v));
-            let round: JobStatus = serde_json::from_str(&json).unwrap();
-            assert_eq!(round, v);
-        }
-    }
-
-    #[test]
-    fn job_status_deserializes_from_lowercase_string() {
-        let s: JobStatus = serde_json::from_str("\"queued\"").unwrap();
-        assert_eq!(s, JobStatus::Queued);
-        let s: JobStatus = serde_json::from_str("\"cancelled\"").unwrap();
-        assert_eq!(s, JobStatus::Cancelled);
-    }
-
-    #[test]
-    fn job_status_rejects_unknown_variant() {
-        assert!(serde_json::from_str::<JobStatus>("\"finished\"").is_err());
-    }
 
     /// Error code string values are part of the wire contract (JSON
     /// output reaches CLI scripts, REST clients, and MCP). Lock the
